Lots of changes:
1. I went back to the project requirements to read through them again
and realized I wasn't following them as I should be. I had to make
a lot of changes to make sure I was using the fac. functions that I
had already written.

2. I am now adding the ship modules to the gameboards so when I
attack a square, and a ship exists on that square, I can then call the
hitShip function on the attacked ship.

3. I changed the naming convention from player & computer to player one
and player two.

4. I moved the show attacks in DOM functionality to a seperate helper
component.

5. I am now checking that each individual ship on the board has been
sunk, and if they have been, updating the winner variable.

I still need to actually write my player turn loop and the AI's attack
logic.