Setup the AI to play against the user
1. I wanted the AI to be smart so if it hit a ship I make it so
it checks the surrounding squares, and continues in a direction if
it continues to hit.

2. I could probably make the ai a lot better. For example, if the
ai hits a ship going left, it will try left again, until it misses. The
problem is the ai doesn't know when it hits a new row. I'm ok with this.
  - ex. ai hits at B1, will try A10 next.

3. There is some temporary logic to reset game, which I will update
tomorrow with the styling.